# all-server-ts/src/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "R2S Backend API",
    "version": "1.0.0",
    "description": "Reserve-to-Save LINE Mini dApp Backend API",
    "contact": {
      "name": "R2S Support",
      "email": "support@r2s.io"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    },
    {
      "url": "https://api.r2s.io",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/campaigns": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "List campaigns",
        "description": "Get list of campaigns with filtering and pagination",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["draft", "recruiting", "reached", "fulfillment", "settled", "failed", "cancelled"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaigns retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Campaigns"],
        "summary": "Create campaign",
        "description": "Create a new campaign",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Campaign": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "chainAddress": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri"
          },
          "merchantWallet": {
            "type": "string"
          },
          "basePrice": {
            "type": "string"
          },
          "minQty": {
            "type": "integer"
          },
          "currentQty": {
            "type": "integer"
          },
          "targetAmount": {
            "type": "string"
          },
          "currentAmount": {
            "type": "string"
          },
          "discountRate": {
            "type": "integer"
          },
          "saveFloorBps": {
            "type": "integer"
          },
          "rMaxBps": {
            "type": "integer"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "recruiting", "reached", "fulfillment", "settled", "failed", "cancelled"]
          },
          "progress": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 100
          }
        }
      },
      "CampaignListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Campaign"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "pages": {
                "type": "integer"
              }
            }
          }
        }
      },
      "CreateCampaignRequest": {
        "type": "object",
        "required": ["title", "basePrice", "minQty", "targetAmount", "discountRate", "startTime", "endTime"],
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri"
          },
          "basePrice": {
            "type": "string"
          },
          "minQty": {
            "type": "integer"
          },
          "targetAmount": {
            "type": "string"
          },
          "discountRate": {
            "type": "integer"
          },
          "saveFloorBps": {
            "type": "integer"
          },
          "rMaxBps": {
            "type": "integer"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}